move desktop notifications into a service
